logic removed from can_delete and tests adapted
